// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/phrase-query-resolution-response-api-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PhraseQueryResolutionResponse } from '../models/phrase-query-resolution-response';
export interface PhraseQueryResolutionResponseApiResponse {
  data: PhraseQueryResolutionResponse | null;
  errors: Array<string> | null;
  isSuccess: boolean;
  message: string | null;
}