// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/phrase-resolution-candidate-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PhraseExactTokenDto } from '../models/phrase-exact-token-dto';
export interface PhraseResolutionCandidateDto {
  displayText: string;
  resolutionRef: string;
  tokens: Array<PhraseExactTokenDto>;
  wordCount: number;
}